requirements.txt
aws-cdk.core
aws-cdk.aws_lambda

requests
